modifing some issues

+++ src/middlewares/rollback-created-documents.middleware.js
@@ -1,6 +1,8 @@
 const rollbackCreatedDocuments = async(req,res,next)=>{
-    const {model , query} = req.createdDocument
-    await model.findByIdAndDelete(query)
+    if(req.createdDocument){
+        const {model , query} = req.createdDocument
+        await model.findByIdAndDelete(query)
+    }
     next()
 }
